requirements.txt
altair==5.5.0
beautifulsoup4==4.13.3
bs4==0.0.1
certifi==2025.1.31
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gspread==6.2.0
lxml==5.3.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
nltk==3.9.1
numpy==2.2.4
oauth2client==4.1.3
pandas==2.2.3
plotly==6.0.1
pyarrow==19.0.1
pymongo==4.6.1
python-dotenv==1.1.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
streamlit==1.43.2
streamlit-aggrid==1.1.2
tabula-py==2.10.0
thinc==8.3.4
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
yarg==0.1.9
yfinance==0.2.55
jpype1
watchdog
python-Levenshtein
sqlalchemy 
pymysql
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp